requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.7.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
